requirements.txt
# Stock Market Automated Trading System - Dependencies (with OpenAlgo)

# OpenAlgo Integration (Infrastructure Layer)
openalgo==1.0.0                 # OpenAlgo Python SDK
requests==2.31.0                # HTTP client for API calls

# Core Libraries
pandas==2.1.4                   # Data manipulation
numpy==1.26.2                   # Numerical computations

# Technical Analysis (Optional - for custom strategies)
pandas-ta==0.3.14b              # Technical indicators
# TA-Lib==0.4.28                # Optional: requires system install

# Configuration
PyYAML==6.0.1                   # YAML configuration files
python-dotenv==1.0.0            # Environment variables

# Database
SQLAlchemy==2.0.23              # ORM for database
alembic==1.13.1                 # Database migrations

# Scheduling
APScheduler==3.10.4             # Task scheduling

# Notifications (Built into OpenAlgo - no need for separate library)
# python-telegram-bot==20.7     # Not needed - OpenAlgo handles this

# Logging & Monitoring
colorlog==6.8.0                 # Colored logging
loguru==0.7.2                   # Advanced logging

# Security (CRITICAL for Financial Applications)
cryptography==44.0.1            # Encryption for credentials
pyotp==2.9.0                    # TOTP for 2FA
pyjwt==2.10.1                   # JWT tokens
argon2-cffi==23.1.0             # Secure password hashing
python-dotenv==1.2.1            # Environment variable management
bcrypt==4.1.3                   # Password hashing

# Utilities
python-dateutil==2.8.2          # Date utilities
pytz==2023.3                    # Timezone handling

# Testing
pytest==7.4.3                   # Unit testing
pytest-cov==4.1.0               # Test coverage
pytest-mock==3.12.0             # Mocking for tests

# Development
black==23.12.1                  # Code formatter
flake8==7.0.0                   # Linting
mypy==1.8.0                     # Type checking

# Data visualization (for backtesting)
matplotlib==3.8.2               # Plotting
seaborn==0.13.0                 # Statistical plots

# Optional: Webhook receiver (if not using OpenAlgo's built-in)
flask==3.0.0                    # Web framework for custom webhooks
flask-cors==4.0.0               # CORS for API

# Optional: Advanced features
scipy==1.11.4                   # Scientific computing
scikit-learn==1.3.2             # Machine learning (for advanced strategies)
